woop: refacotring -> move support to ./components
