+++ client/signup/steps/woocommerce-install/confirm/index.tsx
@@ -1,9 +1,6 @@
 import { NextButton } from '@automattic/onboarding';
 import styled from '@emotion/styled';
-import { createInterpolateElement } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
 import { useI18n } from '@wordpress/react-i18n';
-import { addQueryArgs } from '@wordpress/url';
 import page from 'page';
 import { ReactElement, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
@@ -14,18 +11,11 @@ import { LoadingEllipsis } from 'calypso/components/loading-ellipsis';
 import WarningCard from 'calypso/components/warning-card';
 import StepWrapper from 'calypso/signup/step-wrapper';
 import { submitSignupStep } from 'calypso/state/signup/progress/actions';
-import { getSiteDomain } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import SupportCard from '../components/support-card';
 import useWooCommerceOnPlansEligibility from '../hooks/use-woop-handling';
 import type { WooCommerceInstallProps } from '../';
 import './style.scss';
-
-export const SupportLinkStyle = styled.a`
-	/* Gray / Gray 100 - have to find the var value for this color */
-	color: #101517 !important;
-	text-decoration: underline;
-	font-weight: bold;
-`;
 
 const ActionSection = styled.div`
 	display: flex;
@@ -48,38 +38,12 @@ const WarningsOrHoldsSection = styled.div`
 	margin-bottom: 40px;
 `;
 
-const SupportLinkContainer = styled.p`
-	@media ( max-width: 320px ) {
-		margin-top: 0px;
-		width: 100%;
-	}
-`;
-
 const StyledNextButton = styled( NextButton )`
 	@media ( max-width: 320px ) {
 		width: 100%;
 		margin-bottom: 20px;
 	}
 `;
-
-export function SupportLink(): ReactElement {
-	const siteId = useSelector( getSelectedSiteId ) as number;
-	const domain = useSelector( ( state ) => getSiteDomain( state, siteId ) );
-
-	return (
-		<SupportLinkContainer>
-			{ createInterpolateElement( __( 'Need help? <a>Contact support</a>' ), {
-				a: (
-					<SupportLinkStyle
-						href={ addQueryArgs( '/help/contact', {
-							redirect_to: `/start/woocommerce-install/confirm?site=${ domain }`,
-						} ) }
-					/>
-				),
-			} ) }
-		</SupportLinkContainer>
-	);
-}
 
 export default function Confirm( props: WooCommerceInstallProps ): ReactElement | null {
 	const { goToStep, isReskinned, headerTitle, headerDescription } = props;
@@ -161,7 +125,7 @@ export default function Confirm( props: WooCommerceInstallProps ): ReactElement 
 					{ getCheckoutContent() }
 					{ getWarningsOrHoldsSection() }
 					<ActionSection>
-						<SupportLink />
+						<SupportCard />
 						<StyledNextButton
 							disabled={ hasBlockers || ! isDataReady }
 							onClick={ () => {
